refactor: layout 컴포넌트 분리, URL 변경
