+++ frontend/src/Routes.tsx
@@ -1,33 +1,10 @@
-import { Outlet } from 'react-router-dom';
+import UserLayout from './layouts/UserLayout';
 
 import JobList from '@/pages/JobList';
 import SpaceList from '@/pages/SpaceList';
 import TaskList from '@/pages/TaskList';
 
-import Header from '@/components/Header';
-import InputModal from '@/components/InputModal';
-
-import useModal from '@/hooks/useModal';
-
-const DefaultLayout = () => {
-  const { isShowModal, closeModal } = useModal(!localStorage.getItem('user'));
-
-  return (
-    <>
-      <Header />
-      {isShowModal && (
-        <InputModal
-          title="비밀번호 입력"
-          detail="해당 공간의 관계자만 접근할 수 있습니다."
-          placeholder="비밀번호를 입력해주세요."
-          buttonText="확인"
-          closeModal={closeModal}
-        />
-      )}
-      <Outlet />
-    </>
-  );
-};
+import DefaultLayout from '@/layouts/DefaultLayout';
 
 const routes = [
   {
@@ -35,7 +12,8 @@ const routes = [
     element: <DefaultLayout />,
     children: [
       {
-        path: ':hostId',
+        path: 'enter/:hostId',
+        element: <UserLayout />,
         children: [
           {
             path: 'spaces',
