+++ frontend/src/layouts/DefaultLayout/index.tsx
@@ -1,0 +1,11 @@
+import { Outlet } from 'react-router-dom';
+
+const DefaultLayout = () => {
+  return (
+    <>
+      <Outlet />
+    </>
+  );
+};
+
+export default DefaultLayout;
