+++ frontend/src/layouts/UserLayout/index.tsx
@@ -1,0 +1,28 @@
+import { Outlet } from 'react-router-dom';
+
+import Header from '@/components/Header';
+import InputModal from '@/components/InputModal';
+
+import useModal from '@/hooks/useModal';
+
+const UserLayout = () => {
+  const { isShowModal, closeModal } = useModal(!localStorage.getItem('user'));
+
+  return (
+    <>
+      <Header />
+      {isShowModal && (
+        <InputModal
+          title="비밀번호 입력"
+          detail="해당 공간의 관계자만 접근할 수 있습니다."
+          placeholder="비밀번호를 입력해주세요."
+          buttonText="확인"
+          closeModal={closeModal}
+        />
+      )}
+      <Outlet />
+    </>
+  );
+};
+
+export default UserLayout;
